Changed the printing out to files some to do with shape stuff

+++ protein.h
@@ -37,7 +37,7 @@ int get_J(double difD, double *nATP, double *nADP, double *nE,
           double *JxE, double *JyE, double *JzE);
 void set_membrane(double (*mem_f)(double x, double y, double z), double mem_A[]);
 void set_insideArr(bool *insideArr);
-bool inside(double xi, double yi, double zi);
+bool inside(int xi, int yi, int zi);
 double find_intersection(const double fXYZ, const double fXYz, const double fXyZ, const double fxYZ,
 		       const double fxyZ, const double fxYz, const double fXyz, const double fxyz,
 		       const double f_minus_C);
